requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anthropic==0.57.1
anyio==4.9.0
asttokens==3.0.0
attrs==25.3.0
Authlib==1.6.0
backoff==2.2.1
beautifulsoup4==4.13.4
browser-use==0.5.4
bubus==1.4.5
cachetools==5.5.2
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
comm==0.2.2
contourpy==1.2.0
cryptography==45.0.5
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
distro==1.9.0
dokusan==0.1.0
dotenv==0.9.9
executing==2.2.0
fastapi==0.115.6
filetype==1.2.0
fonttools==4.49.0
google==3.0.0
google-ai-generativelanguage==0.6.18
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-genai==1.25.0
googleapis-common-protos==1.70.0
greenlet==3.2.3
groq==0.30.0
grpcio==1.73.1
grpcio-status==1.73.1
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
ipykernel==6.29.5
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.5
langchain==0.3.26
langchain-core==0.3.68
langchain-google-genai==2.1.7
langchain-groq==0.3.6
langchain-text-splitters==0.3.8
langgraph==0.5.2
langgraph-checkpoint==2.1.0
langgraph-prebuilt==0.5.2
langgraph-sdk==0.1.72
langsmith==0.4.5
markdown-it-py==3.0.0
markdown_pdf==1.7
markdownify==1.1.0
matplotlib==3.8.3
matplotlib-inline==0.1.7
mcp==1.11.0
mdurl==0.1.2
MouseInfo==0.1.3
nest-asyncio==1.6.0
node==1.2.2
npm==0.1.1
npx==0.1.6
numpy==1.26.4
oauthlib==3.3.1
odict==1.9.0
ollama==0.5.1
openai==1.95.1
optional-django==0.1.0
orjson==3.10.18
ormsgpack==1.10.0
packaging==23.2
pandas==2.2.1
parso==0.8.4
patchright==1.52.5
pillow==10.2.0
platformdirs==4.3.8
playwright==1.53.0
plumber==1.7
portalocker==2.10.1
posthog==6.1.0
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==6.31.1
psutil==7.0.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAutoGUI==0.9.54
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pyee==13.0.0
PyGetWindow==0.0.9
Pygments==2.19.2
PyMsgBox==1.0.9
PyMuPDF==1.25.3
pyparsing==3.1.2
pypdf==5.7.0
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==0.1.30
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
python-telegram-bot==22.2
pytweening==1.2.0
pytz==2024.1
pywin32==310
PyYAML==6.0.2
pyzmq==27.0.0
referencing==0.36.2
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rpds-py==0.26.0
rsa==4.9.1
screeninfo==0.8.1
setuptools==74.0.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
sse-starlette==2.4.1
stack-data==0.6.3
starlette==0.41.3
tenacity==8.5.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.2.0
urllib3==2.5.0
uuid7==0.1.0
uvicorn==0.35.0
wcwidth==0.2.13
websockets==15.0.1
xxhash==3.5.0
zope.component==6.0
zope.deferredimport==5.0
zope.deprecation==5.0
zope.event==5.0
zope.hookable==6.0
zope.interface==7.0.3
zope.lifecycleevent==5.0
zope.proxy==5.3
zstandard==0.23.0
